+++ src/main/java/com/almostreliable/summoningrituals/core/Registration.java
@@ -4,16 +4,21 @@ import com.almostreliable.summoningrituals.ModConstants;
 import com.almostreliable.summoningrituals.altar.AltarBlock;
 import com.almostreliable.summoningrituals.altar.AltarBlockEntity;
 import com.almostreliable.summoningrituals.data.SummoningLang;
+import com.almostreliable.summoningrituals.dev.PatternGeneratorItem;
+import com.almostreliable.summoningrituals.dev.PatternGeneratorState;
 import com.almostreliable.summoningrituals.recipe.AltarRecipe;
 import com.almostreliable.summoningrituals.recipe.AltarRecipeSerializer;
 import com.almostreliable.summoningrituals.recipe.condition.ConditionRegistry;
 import com.almostreliable.summoningrituals.recipe.condition.check.AltarTagBlockStateCheck;
 import com.almostreliable.summoningrituals.recipe.condition.check.MoonPhaseCheck;
 
+import net.minecraft.core.component.DataComponentType;
+import net.minecraft.core.component.DataComponents;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.CreativeModeTab;
-import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.Rarity;
 import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.level.block.Block;
@@ -30,9 +35,10 @@ import net.neoforged.neoforge.capabilities.Capabilities;
 import net.neoforged.neoforge.capabilities.RegisterCapabilitiesEvent;
 import net.neoforged.neoforge.registries.DeferredBlock;
 import net.neoforged.neoforge.registries.DeferredHolder;
+import net.neoforged.neoforge.registries.DeferredItem;
 import net.neoforged.neoforge.registries.DeferredRegister;
 
-import java.util.Collection;
+import java.util.List;
 import java.util.function.Function;
 
 @SuppressWarnings("ConstantConditions")
@@ -48,10 +54,14 @@ public final class Registration {
     private static final DeferredRegister<RecipeType<?>> RECIPE_TYPES = DeferredRegister.create(Registries.RECIPE_TYPE, ModConstants.MOD_ID);
     private static final DeferredRegister<RecipeSerializer<?>> RECIPE_SERIALIZERS = DeferredRegister.create(Registries.RECIPE_SERIALIZER, ModConstants.MOD_ID);
     private static final DeferredRegister<LootItemConditionType> LOOT_CONDITIONS = DeferredRegister.create(Registries.LOOT_CONDITION_TYPE, ModConstants.MOD_ID);
+    private static final DeferredRegister.DataComponents DATA_COMPONENTS = DeferredRegister.createDataComponents(Registries.DATA_COMPONENT_TYPE, ModConstants.MOD_ID);
 
     // blocks
     public static final DeferredBlock<AltarBlock> ALTAR_BLOCK = registerBlock(Constants.ALTAR, "Summoning Altar", AltarBlock::new, p -> p.strength(2.5f));
     public static final DeferredBlock<AltarBlock> INDESTRUCTIBLE_ALTAR_BLOCK = registerBlock(Constants.INDESTRUCTIBLE_ALTAR, "Indestructible Summoning Altar", AltarBlock::new, p -> p.strength(-1.0f, 3_600_000.0f));
+
+    // items
+    public static final DeferredItem<PatternGeneratorItem> PATTERN_GENERATOR_ITEM = ITEMS.registerItem("pattern_generator", PatternGeneratorItem::new, new Item.Properties().stacksTo(1).rarity(Rarity.EPIC).component(DataComponents.ENCHANTMENT_GLINT_OVERRIDE, true));
 
     // block entities
     public static final DeferredHolder<BlockEntityType<?>, BlockEntityType<AltarBlockEntity>> ALTAR_BLOCK_ENTITY = registerBlockEntity(ALTAR_BLOCK.getId(), AltarBlockEntity::new, ALTAR_BLOCK, INDESTRUCTIBLE_ALTAR_BLOCK);
@@ -71,13 +81,19 @@ public final class Registration {
     public static final DeferredHolder<LootItemConditionType, LootItemConditionType> ALTAR_TAG_BLOCK_STATE_CONDITION = LOOT_CONDITIONS.register(Constants.BLOCK_TAG_BLOCK_STATE, () -> new LootItemConditionType(AltarTagBlockStateCheck.CODEC));
     public static final DeferredHolder<LootItemConditionType, LootItemConditionType> MOON_PHASE_CONDITION = LOOT_CONDITIONS.register(Constants.MOON_PHASE, () -> new LootItemConditionType(MoonPhaseCheck.CODEC));
 
+    // data components
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<PatternGeneratorState>> PATTERN_GENERATOR_COMPONENT = DATA_COMPONENTS.registerComponentType(
+        "pattern_generator_state",
+        builder -> builder.networkSynchronized(PatternGeneratorState.STREAM_CODEC)
+    );
+
     // creative tab
     public static final DeferredHolder<CreativeModeTab, CreativeModeTab> TAB = CREATIVE_TABS.register(
         "tab", () -> CreativeModeTab.builder()
             .title(SummoningLang.LangEntry.of("tab", "main", ModConstants.MOD_NAME).get())
             .icon(ALTAR_BLOCK::toStack)
             .noScrollBar()
-            .displayItems((features, output) -> output.acceptAll(getKnownItems()))
+            .displayItems((features, output) -> output.acceptAll(List.of(ALTAR_BLOCK.toStack(), INDESTRUCTIBLE_ALTAR_BLOCK.toStack())))
             .build()
     );
 
@@ -93,13 +109,10 @@ public final class Registration {
         RECIPE_TYPES.register(eventBus);
         RECIPE_SERIALIZERS.register(eventBus);
         LOOT_CONDITIONS.register(eventBus);
+        DATA_COMPONENTS.register(eventBus);
 
         eventBus.addListener(Registration::onCommonSetup);
         eventBus.addListener(Registration::registerCapabilities);
-    }
-
-    private static Collection<ItemStack> getKnownItems() {
-        return ITEMS.getEntries().stream().map(e -> e.value().getDefaultInstance()).toList();
     }
 
     private static void onCommonSetup(FMLCommonSetupEvent event) {