+++ src/main/java/com/almostreliable/summoningrituals/dev/package-info.java
@@ -1,0 +1,4 @@
+@NonnullDefault
+package com.almostreliable.summoningrituals.dev;
+
+import org.lwjgl.system.NonnullDefault;