+++ src/main/java/com/almostreliable/summoningrituals/network/HighlightPositionsPacket.java
@@ -31,6 +31,6 @@ public record HighlightPositionsPacket(List<BlockPos> positions) implements Cust
         if (level == null) return;
 
         var gameTime = level.getGameTime();
-        BlockHighlightRenderer.highlightPositions(gameTime, packet.positions);
+        BlockHighlightRenderer.highlightWrongPositions(gameTime, packet.positions);
     }
 }